Migrate to HandyControl 2.1.0
Use the official version istead of a fork with tweaks. The tweaks should
be moved to MovieList directly so that there wouldn't be a need for a
fork of HandyControl.

This commit doesn't include all the changes required for the app to look
like it looked before (or like it should look in the future).