Create the initial logging configuration
Configure the Locator, including view location, database services (of
which there are none at the moment), and logging. Use Serilog for
logging and log to a file and the debug window.

The logging confiugration created in this commit is simple and will be
changed to get the configuration from a file.

 #12