Implement opening files through the home page
Implement opening the file dialog and opening a tab when the file is
selected.

Also, implement shifting the selected tab when a file is closed. The
implementation is not perfect and should be refactored.

Downgrade to Material Design 2.6.0, because Material Design Extensions
requires this version to open file dialogs.

Change the UI language to English.

 #18