Refactor opening files
Also, fix a weird bug with opening files from the recent files list.
When one file is opened and an entry in the recent files list is
double-clicked, the event is fired twice. This happens only when a file
is already opened. If a file is nonexistent the app tries to open a
dialog twice, which throws an exception.

The fix just makes the dialog interaction exit if a dialog is already
opened which doesn't address the underlying issue of the event firing
twice. Maybe the underlying issue will be resolved in a future commit.