Implement the list
Implement the list of movies and series, which is the single most
important feature of the app.

Remove some CHECK constaints from the database schema for files, because
they weren't working correctly.

The list header for Year has the length 142 instead of 125, as with
columns for Year in the list. This is to account for the scroll bar.
(ListControl.xaml:28)

 #19