Implement deleting a movie
Create a movie service, which can save and delete movies (saving is not
implemented yet). Delete related entities when a movie is deleted.

Implement removing the list item when a movie is deleted and updating
related items.

Add attributes to each entity so that Dapper can update them properly.

Refactor the movie form control code-behind.

 #20