Implement showing seasons in the series form
Show the list of seasons in the series form. Showing the season form
when a list item is clicked is not implemented yet.

Do some refactorings in the form view models. Change the period overlap
rule to a simple observable of bools in the season form view model
because of a weird NullReferenceException.

 #21