Remove support for INotifyDataErrorInfo from reactive forms
The red border in the period form was not actually fixed previously. Fix
it by always returning an empty sequence of errors (even if they are
present) from the GetErrors method. This doesn't seem to influence the
actual validation process, but maybe a better way of doing this will be
found in the future.