Add FluentAvalonia
FluentAvalonia provies lots of custom controls and a more modern variant
of the Fluent UI even though it's a little rough around the edges.

This commit only includes this library and introduces minor changes to
the UI, most notably changing the tab control in the file view to a
navigation pane, which means that apparently no changes will be needed
in the view model. More changes will come in the following commits as
the UI code will need to be overhauled to use this library everywhere.

 #68