Refactor and fix validation in the app
Most validation rules don't need to override the first error message
since they are not invalid in new forms. Only a couple rules need this.

For some absolutely inexplicable reason, some message overrides work,
and some don't, and I couldn't for the life of me find out why. That's
why the BindStrictValidation contains the 'magicallyWorks' parameter -
if it's true it means that overriding the error message only once will
work as expected. If it doesn't (like for the movie or series forms)
then it needs to override the error message 4 times. Why 4? Hell if I
know. But now it works and I'm currenly OK with this hack.