Improve the usability of number boxes
Make number boxes across the app writable. The secret to changing their
value manually was pressing 'Enter' all along, so add a tooltip which
explains that. Also, add a converter from double to int which can handle
NaN (it occurs when the number box is empty).

Also, when a value is bound to a number box, it requests to bring it
into view for some reason. This is absurd. Disable this audacity by
always marking this event as handled in movie and period forms.