Remove WAL from SQLite connections
Write-ahead logging is cool and all, but it means that two temporary
files need to be managed for every list file which is not ideal since
every list file should be self-contained. The performance difference of
various journal modes doesn't matter much since the transactions are
usually very small and don't happen concurrently.