Remove everything and start anew... again
Building a desktop app for tracking movies and series was certainly fun
and I've learned a lot. But the time has come to make this a web app
since desktop apps fall out of fashion for such things. Also, this will
make the app truly cross-platform (Avalonia surely makes it easy, but
nothing beats a browser) and I'll be able to use it on mobile as well
without writing native apps.

After creating the project skeleton, some logic will be moved from the
previous version, but there'll be a lot of new stuff as well.