Rewrite the entire front-end validation framework
Now instead of validating fields separately, a request model is created
on every change, and all validation errors are captured at once. The
validation error messages will be shown on fields based on the change of
their triggers' values.

Previously I wanted to execute validators on requests without creating
those requests, but it proved to not be possible. So now I've done it
the opposite way - the request doeasn't need to be created when the form
is saved - it's already there if the form is valid.

GH-98