Sort the list on the back end
Add a new table just for sorting list items and store sortable titles
which take franchises into account. Currently, some functionality is
broken, e.g. updating the list without reloading the page after adding
or removing a movie or series, but it will be fixed when pagination is
added.

GH-112