Implement the basics of logging in correctly
Implementing authentication and authorization required actually using
Blazor render mode and only use WebAssembly without prerendering (which
enables using Fluxor) only in the main application pages. This allowed
using static rendering for the login page which doesn't actually
requires any interactivity and can thus execute code on the server.

This is only the basics of logging in as multiple features are still not
done - only the main happy path is.

GH-115